+++ src/features/multi-profile/ArchiveProfile.tsx
@@ -5,6 +5,10 @@ import { View } from 'native-base';
 import i18n from '@covid/locale/i18n';
 import { RegularBoldText, RegularText } from '@covid/components/Text';
 import { colors } from '@theme';
+import { ArchiveConfirmationModal } from '@covid/features/multi-profile/ArchiveConfirmationModal';
+import { ArchiveReasonModal } from '@covid/features/multi-profile/ArchiveReasonModal';
+import Navigator from '@covid/features/Navigation';
+import UserService from '@covid/core/user/UserService';
 
 type Props = {
   profileId: string;
@@ -14,119 +18,44 @@ export const ArchiveProfile: React.FC<Props> = (props) => {
   const [isConfirmModalVisible, setConfirmModalVisibility] = useState(false);
   const [isReasonModalVisible, setReasonModalVisibility] = useState(false);
 
-  function handleArchive() {
+  function clickArchive() {
     setConfirmModalVisibility(true);
+  }
+
+  function cancelArchive() {
+    setConfirmModalVisibility(false);
+  }
+
+  function confirmArchive() {
+    setConfirmModalVisibility(false);
+    setReasonModalVisibility(true);
+  }
+
+  function submitReason(reason: string) {
+    setReasonModalVisibility(false);
+
+    const infos = {
+      archived: true,
+      archived_reason: reason,
+    };
+
+    const userService = new UserService();
+    userService.updatePatient(props.profileId, infos).then((response) => {
+      Navigator.gotoScreen('SelectProfile');
+    });
   }
 
   return (
     <>
-      <View style={styles.centeredView}>
-        <Modal transparent visible={isConfirmModalVisible}>
-          <View style={styles.centeredView}>
-            <View style={styles.modalView}>
-              <RegularBoldText style={styles.ratingHeader}>Archive this profile?</RegularBoldText>
-              <RegularText style={styles.ratingText}>
-                You will no longer be able on behalf of this individual. Do you want to proceed?
-              </RegularText>
+      {isConfirmModalVisible && (
+        <ArchiveConfirmationModal cancelArchive={cancelArchive} confirmArchive={confirmArchive} />
+      )}
 
-              <View style={styles.actionContainer}>
-                <TouchableOpacity style={styles.ratingButton} onPress={() => setConfirmModalVisibility(false)}>
-                  <RegularText style={styles.buttonText}>Cancel</RegularText>
-                </TouchableOpacity>
-                <View style={styles.verticalDivider} />
-                <TouchableOpacity
-                  style={styles.ratingButton}
-                  onPress={() => {
-                    setConfirmModalVisibility(false);
-                    setReasonModalVisibility(true);
-                  }}>
-                  <RegularText style={styles.buttonText}>Archive</RegularText>
-                </TouchableOpacity>
-              </View>
-            </View>
-          </View>
-        </Modal>
-      </View>
+      {isReasonModalVisible && <ArchiveReasonModal submitReason={submitReason} />}
 
-      <View style={styles.centeredView}>
-        <Modal transparent visible={isReasonModalVisible} onRequestClose={() => {}}>
-          <View style={styles.centeredView}>
-            <View style={styles.modalView}>
-              <RegularBoldText style={styles.ratingHeader}>Profile Archived</RegularBoldText>
-              <RegularText style={styles.ratingText}>
-                Help us understand why you archived this profile, so we can more accurately predict levels of COVID
-              </RegularText>
-
-              <View style={styles.actionContainer}>
-                <TouchableOpacity style={styles.ratingButton} onPress={() => setReasonModalVisibility(false)}>
-                  <RegularText style={styles.buttonText}>Cancel</RegularText>
-                </TouchableOpacity>
-              </View>
-            </View>
-          </View>
-        </Modal>
-      </View>
-
-      <TouchableOpacity onPress={() => handleArchive()}>
+      <TouchableOpacity onPress={() => clickArchive()}>
         <RegularText>{i18n.t('edit-profile.archive-cta')}</RegularText>
       </TouchableOpacity>
     </>
   );
 };
-
-const actionButtonBorder = 'rgba(240, 240, 240, 1)';
-const styles = StyleSheet.create({
-  centeredView: {
-    backgroundColor: 'rgba(0,0,0,0.6)',
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  modalView: {
-    margin: 30,
-    backgroundColor: 'white',
-    borderRadius: 20,
-    paddingTop: 35,
-    alignItems: 'center',
-    shadowColor: colors.black,
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
-  },
-  verticalDivider: {
-    height: '100%',
-    width: 1,
-    backgroundColor: actionButtonBorder,
-  },
-  ratingText: {
-    paddingBottom: 30,
-    marginHorizontal: 60,
-    fontSize: 14,
-    textAlign: 'center',
-  },
-  ratingHeader: {
-    paddingBottom: 10,
-    fontSize: 18,
-    textAlign: 'center',
-  },
-  actionContainer: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    borderTopWidth: 1,
-    borderColor: actionButtonBorder,
-  },
-  buttonText: {
-    textAlign: 'center',
-    color: colors.linkBlue,
-  },
-  ratingButton: {
-    width: '50%',
-    height: 60,
-    justifyContent: 'center',
-    color: colors.linkBlue,
-  },
-});
